refactor: remove old CourseView files and rename NewCourseView directory
