+++ apps/web/app/modules/Courses/CourseView/CourseOverview.tsx
