+++ apps/web/app/modules/Courses/CourseView/CourseView.page.tsx
@@ -1,101 +1,56 @@
-import {
-  type ClientLoaderFunctionArgs,
-  isRouteErrorResponse,
-  useParams,
-  useRouteError,
-} from "@remix-run/react";
+import { useParams } from "@remix-run/react";
 
-import { courseQueryOptions, useCourseSuspense } from "~/api/queries";
-import { useScormMetadata } from "~/api/queries/admin/useScormMetaData";
-import { queryClient } from "~/api/queryClient";
+import { useCourse } from "~/api/queries";
 import { PageWrapper } from "~/components/PageWrapper";
-import {
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbList,
-  BreadcrumbSeparator,
-} from "~/components/ui/breadcrumb";
-import { useUserRole } from "~/hooks/useUserRole";
-import { cn } from "~/lib/utils";
-import CustomErrorBoundary from "~/modules/common/ErrorBoundary/ErrorBoundary";
+import { CourseChapter } from "~/modules/Courses/CourseView/CourseChapter";
+import CourseOverview from "~/modules/Courses/CourseView/CourseOverview";
+import { CourseViewSidebar } from "~/modules/Courses/CourseView/CourseViewSidebar/CourseViewSidebar";
+import { MoreCoursesByAuthor } from "~/modules/Courses/CourseView/MoreCoursesByAuthor";
+import { YouMayBeInterestedIn } from "~/modules/Courses/CourseView/YouMayBeInterestedIn";
 
-import { CourseViewMainCard } from "./CourseViewMainCard";
-import { LessonsList } from "./LessonsList";
+export default function CourseViewPage() {
+  const { id = "" } = useParams();
+  const { data: course } = useCourse(id);
 
-import type { MetaFunction } from "@remix-run/node";
+  if (!course) return null;
 
-export const meta: MetaFunction = () => {
-  return [{ title: "Courses" }, { name: "description", content: "Courses" }];
-};
-
-export const clientLoader = async ({ params }: ClientLoaderFunctionArgs) => {
-  const courseId = params.id;
-  if (!courseId) return null;
-  await queryClient.prefetchQuery(courseQueryOptions(courseId));
-  return null;
-};
-
-export default function CoursesViewPage() {
-  const { isAdmin } = useUserRole();
-  const { id } = useParams<{ id: string }>();
-
-  const { data: course } = useCourseSuspense(id ?? "");
-  const { data: scormMetadata } = useScormMetadata({
-    id: id ?? "",
-    isScorm: course?.isScorm,
-  });
-
-  if (!id) {
-    throw new Error("No course ID provided");
-  }
-
-  if (!course) {
-    throw new Error("Course not found");
-  }
+  // TODO: Add breadcrumbs
+  // const breadcrumbs = [
+  //   {
+  //     title: "Dashboard",
+  //     href: "/",
+  //   },
+  //   {
+  //     title: course?.title ?? "",
+  //     href: `/course/${id}`,
+  //   },
+  // ];
 
   return (
-    <PageWrapper
-      className={cn("flex flex-col gap-y-4 md:gap-y-6", {
-        "h-full": course.isScorm,
-      })}
-    >
-      <BreadcrumbList>
-        <BreadcrumbItem>
-          <BreadcrumbLink href="/">Dashboard</BreadcrumbLink>
-        </BreadcrumbItem>
-        <BreadcrumbSeparator />
-        <BreadcrumbItem>
-          <BreadcrumbLink href={`/course/${id}`}>{course.title}</BreadcrumbLink>
-        </BreadcrumbItem>
-      </BreadcrumbList>
-      <div className="flex flex-col md:flex-row h-full gap-6">
-        {/* {course.isScorm ? (
-          <iframe
-            title={scormMetadata?.entryPoint}
-            src={`/api/scorm/${id}/content?path=${scormMetadata?.entryPoint}`}
-            width="100%"
-            height="100%"
-            className="w-full h-full"
-          ></iframe>
-        ) : ( */}
-        <>
-          <CourseViewMainCard {...course} />
-          <LessonsList lessons={course.chapters} isEnrolled={course.enrolled || isAdmin} />
-        </>
-        {/* )} */}
+    <PageWrapper className="max-w-full">
+      <div className="flex flex-col lg:grid lg:grid-cols-[1fr_480px] max-w-full gap-6 w-full">
+        <div className="flex flex-col gap-y-6 overflow-hidden">
+          <CourseOverview course={course} />
+          <div className="py-6 px-4 md:p-8 rounded-lg bg-white flex flex-col gap-y-4">
+            <div className="flex flex-col gap-y-1">
+              <h4 className="h6 text-neutral-950">Chapters</h4>
+              <p className="body-base-md text-neutral-800">
+                Below you can see all chapters inside this course
+              </p>
+            </div>
+            {course?.chapters?.map((chapter) => {
+              if (!chapter) return null;
+
+              return (
+                <CourseChapter chapter={chapter} key={chapter.id} enrolled={course.enrolled} />
+              );
+            })}
+          </div>
+          <MoreCoursesByAuthor courseId={course.id} teacherId={course.authorId} />
+          <YouMayBeInterestedIn courseId={course.id} category={course.category} />
+        </div>
+        <CourseViewSidebar course={course} />
       </div>
     </PageWrapper>
   );
-}
-
-export function ErrorBoundary() {
-  const error = useRouteError();
-
-  if (isRouteErrorResponse(error)) {
-    return <CustomErrorBoundary stack={error.data} />;
-  } else if (error instanceof Error) {
-    return <CustomErrorBoundary stack={error.stack} />;
-  } else {
-    return <CustomErrorBoundary />;
-  }
 }
