+++ apps/web/app/modules/Courses/CourseView/CourseViewSidebar/CourseOptions.tsx
