+++ apps/web/app/modules/Courses/CourseView/MoreCoursesByAuthor.tsx
