--- apps/web/app/modules/Courses/CourseView/UnenrollDialog.tsx
@@ -1,52 +1,0 @@
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "~/components/ui/alert-dialog";
-import { Button } from "~/components/ui/button";
-
-export default function UnenrollAlertDialog({
-  setIsEnrolled,
-  buttonText,
-  title,
-  description,
-}: {
-  setIsEnrolled: React.Dispatch<React.SetStateAction<boolean>>;
-  buttonText: string;
-  title: string;
-  description: string;
-}) {
-  return (
-    <AlertDialog>
-      <AlertDialogTrigger asChild>
-        <Button className="mt-2 w-full border bg-white border-secondary-500 text-secondary-700 py-2 rounded-lg">
-          {buttonText}
-        </Button>
-      </AlertDialogTrigger>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>{title}</AlertDialogTitle>
-          <AlertDialogDescription>{description}</AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter className="sm:justify-start">
-          <AlertDialogAction asChild>
-            <Button className="bg-destructive" onClick={() => setIsEnrolled(false)}>
-              Unenroll
-            </Button>
-          </AlertDialogAction>
-          <AlertDialogCancel asChild>
-            <Button variant="outline" className="text-primary-800">
-              Cancel
-            </Button>
-          </AlertDialogCancel>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
