+++ apps/web/app/modules/Courses/CourseView/YouMayBeInterestedIn.tsx
