+++ apps/web/app/modules/Courses/CourseView/components/ChapterCounter.tsx
