+++ apps/web/app/modules/Courses/CourseView/components/CourseProgressChart.tsx
