+++ apps/web/app/modules/Courses/CourseView/index.ts
@@ -1,0 +1,1 @@
+export { CourseViewPage } from "./CourseView.page";
