+++ apps/web/app/modules/Courses/CourseView/lessonTypes.ts
