--- apps/web/app/modules/Courses/NewCourseView/index.ts
@@ -1,1 +1,0 @@
-export { NewCourseViewPage } from "./NewCourseView.page";
